requirements.txt
AccessControl==5.0
Acquisition==4.7
alembic==1.5.5
amqp==5.0.5
aniso8601==9.0.0
appnope==0.1.2
arrow==0.17.0
attrs==20.3.0
AuthEncoding==4.2
b2handle==1.1.2
Babel==2.9.0
backcall==0.2.0
base32-lib==1.0.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.3.0
blinker==1.4
BTrees==4.7.2
build==0.3.0
cachelib==0.1.1
cchardet==2.1.7
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
Chameleon==3.9.0
chardet==4.0.0
check-manifest==0.46
click==7.1.2
click-default-group==1.2.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.4
cryptography==3.4.6
datacite==1.0.1
DateTime==4.3
dcxml==0.1.2
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
dnspython==2.1.0
Docker-Services-CLI==0.3.0
DocumentTemplate==4.0
dojson==1.4.0
doschema==1.0.0a1
easywebdav2==1.3.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
email-validator==1.1.2
entrypoints==0.3
ExtensionClass==4.5.0
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Alembic==2.0.1
Flask-BabelEx==0.9.4
Flask-Breadcrumbs==0.5.1
Flask-Caching==1.9.0
Flask-CeleryExt==0.3.4
Flask-Collect==1.2.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-IIIF==0.6.1
Flask-KVSession-Invenio==0.6.3
Flask-Limiter==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Menu==0.7.2
Flask-OAuthlib==0.9.6
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security==3.0.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
flask-talisman==0.5.0
Flask-Testing==0.8.1
flask-webpackext==1.0.2
Flask-WTF==0.14.3
fs==0.5.4
ftfy==4.4.3
future==0.18.2
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==3.7.0
importmagic3==0.2.0
infinity==1.5
iniconfig==1.1.1
intervals==0.9.1
invenio==3.3.0
invenio-access==1.4.2
invenio-accounts==1.4.3
invenio-accounts-rest==1.0.0a4
invenio-admin==1.3.0
invenio-app==1.2.7
invenio-assets==1.2.5
invenio-base==1.2.3
invenio-cache==1.1.0
invenio-celery==1.2.2
invenio-config==1.0.3
invenio-db==1.0.8
invenio-deposit==1.0.0a11
invenio-files-rest==1.2.0
invenio-formatter==1.1.0
invenio-i18n==1.2.0
invenio-indexer==1.2.0
invenio-jsonschemas==1.1.1
invenio-mail==1.0.2
invenio-marc21==1.0.0a9
invenio-oaiserver==1.2.0
invenio-oauth2server==1.2.0
invenio-oauthclient==1.4.4
invenio-pidrelations==1.0.0a4
invenio-pidstore==1.2.2
invenio-previewer==1.2.1
invenio-queues==1.0.0a3
invenio-records==1.4.0
invenio-records-files==1.2.1
invenio-records-rest==1.7.2
invenio-records-ui==1.1.0
invenio-rest==1.2.3
invenio-search==1.4.1
invenio-search-ui==2.0.2
invenio-stats==1.0.0a18 
invenio-theme==1.3.8
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsmin==2.2.2
jsonpatch==1.28
jsonpointer==2.0
jsonref==0.2
jsonresolver==0.3.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kombu==5.0.2
limits==1.5.1
log-colorizer==1.8.6
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
maxminddb==2.0.3
maxminddb-geolite2==2018.703
mistune==0.8.4
mock==4.0.3
msgpack==1.0.2
MultiMapping==4.1
multipart==0.2.4
nbconvert==5.6.1
nbformat==5.1.2
node-semver==0.1.1
oauthlib==2.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
passlib==1.7.4
PasteDeploy==2.1.1
pep517==0.9.1
Persistence==3.0
persistent==4.6.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.3
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
Pygments==2.8.0
PyJWT==2.0.1
pynpm==0.1.2
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
pytest-dotenv==0.5.2
pytest-flask==1.1.0
pytest-invenio==1.4.1
pytest-isort==1.3.0
pytest-mock==3.5.1
pytest-pycodestyle==2.2.0
pytest-pydocstyle==2.2.0
pytest-runner==5.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-gettext==4.0
python-magic==0.4.20
pytz==2021.1
pywebpack==1.2.0
pyzmq==22.0.3
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.1.0
responses==0.12.1
RestrictedPython==5.1
roman==3.3
selenium==3.141.0
simplejson==3.17.2
simplekv==0.14.1
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2
spark-parser==1.8.9
speaklater==1.3
SQLAlchemy==1.3.23
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.35.0
testpath==0.4.4
toml==0.10.2
tornado==5.1.1
tornado-systemd==1.0.1
traitlets==4.3.3
transaction==3.0.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
uncompyle6==3.7.4
uritools==3.0.0
urllib3==1.26.3
validators==0.18.2
vine==5.0.0
waitress==1.4.4
Wand==0.6.5
wcwidth==0.2.5
wdb==3.3.0
wdb.server==3.3.0
webargs==5.5.3
webencodings==0.5.1
WebOb==1.8.7
WebTest==2.0.35
Werkzeug==1.0.1
WSGIProxy2==0.4.6
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
xdis==5.0.7
z3c.pt==3.3.0
zc.lockfile==2.0
ZConfig==3.5.0
zExceptions==4.1
zipp==3.4.0
ZODB==5.6.0
zodbpickle==2.0.0
Zope==5.1.1
zope.annotation==4.7.0
zope.browser==2.3
zope.browsermenu==4.4
zope.browserpage==4.4.0
zope.browserresource==4.4
zope.cachedescriptors==4.3.1
zope.component==4.6.2
zope.configuration==4.4.0
zope.container==4.4.0
zope.contentprovider==4.2.1
zope.contenttype==4.5.0
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.dottedname==4.3
zope.event==4.5.0
zope.exceptions==4.4
zope.filerepresentation==5.0.0
zope.globalrequest==1.5
zope.hookable==5.0.1
zope.i18n==4.7.0
zope.i18nmessageid==5.0.1
zope.interface==5.2.0
zope.lifecycleevent==4.3
zope.location==4.2
zope.pagetemplate==4.5.0
zope.processlifetime==2.3.0
zope.proxy==4.3.5
zope.ptresource==4.2.0
zope.publisher==6.0.0
zope.schema==6.1.0
zope.security==5.1.1
zope.sequencesort==4.1.2
zope.site==4.4.0
zope.size==4.3
zope.structuredtext==4.3
zope.tal==4.4
zope.tales==5.1
zope.testbrowser==5.5.1
zope.testing==4.9
zope.traversing==4.4.1
zope.viewlet==4.2.1
